refactor Module.cpp
