+++ torch_npu/csrc/npu/Module.cpp
@@ -21,7 +21,6 @@
 #include "csrc/npu/NPUCachingAllocator.h"
 #include "npu/core/NPUException.h"
 #include "csrc/npu/NPUFunctions.h"
-#include "npu/core/NPUGuard.h"
 #include "csrc/npu/NPUStream.h"
 #include "npu/core/NpuVariables.h"
 #include "npu/core/register/OptionRegister.h"
@@ -29,94 +28,6 @@
 #include "torch_npu/csrc/npu/Module.h"
 #include "torch_npu/csrc/npu/NPUPluggableAllocator.h"
 #include "torch_npu/csrc/npu/memory_snapshot.h"
-
-struct NPUDeviceProp {
-  std::string name;
-  size_t totalGlobalMem = 0;
-};
-
-struct NPUDeviceMem {
-  size_t totalGlobalMem = 0;
-  size_t freeMem = 0;
-};
-NPUDeviceProp prop;
-void RegisterNPUDeviceProperties(PyObject* module) {
-  auto m = py::handle(module).cast<py::module>();
-  py::class_<NPUDeviceProp>(m, "_NPUDeviceProperties")
-      .def_readonly("name", &NPUDeviceProp::name)
-      .def_readonly("total_memory", &NPUDeviceProp::totalGlobalMem)
-      .def("__repr__", [](const NPUDeviceProp& prop) {
-        std::ostringstream stream;
-        stream << "_NPUDeviceProperties(name='" << prop.name
-               << "', total_memory="
-               << prop.totalGlobalMem / (CHANGE_UNIT_SIZE * CHANGE_UNIT_SIZE)
-               << "MB)";
-        return stream.str();
-      });
-  m.def(
-      "_npu_record_memory_history",
-      static_cast<void (*)(
-          c10::optional<std::string>,
-          c10::optional<std::string>,
-          std::string,
-          size_t)>(torch_npu::_record_memory_history));
-
-  m.def("_npu_isHistoryEnabled", []() {
-    return c10_npu::NPUCachingAllocator::isHistoryEnabled();
-  });
-}
-
-NPUDeviceProp* GetDeviceProperties(int64_t deviceid) {
-  const char* device_name;
-  size_t device_free;
-  size_t device_total;
-  device_name = c10_npu::acl::AclrtGetSocName();
-  if (device_name == nullptr) {
-    prop.name = " ";
-    ASCEND_LOGE("NPU get device name fail.");
-  } else {
-    prop.name = std::string(device_name);
-  }
-  NPU_CHECK_ERROR(aclrtGetMemInfo(ACL_HBM_MEM, &device_free, &device_total));
-  prop.totalGlobalMem = device_total;
-  return &prop;
-}
-
-void BindGetDeviceProperties(PyObject* module) {
-  auto m = py::handle(module).cast<py::module>();
-  m.def(
-      "_npu_getDeviceProperties",
-      [](int deviceid) -> NPUDeviceProp* {
-        return GetDeviceProperties(deviceid);
-      },
-      py::return_value_policy::reference);
-}
-
-NPUDeviceMem memory;
-void RegisterNPUDeviceMemories(PyObject* module) {
-  auto m = py::handle(module).cast<py::module>();
-  py::class_<NPUDeviceMem>(m, "_NPUDeviceMemories")
-      .def_readonly("total_memory", &NPUDeviceMem::totalGlobalMem)
-      .def_readonly("free_memory", &NPUDeviceMem::freeMem);
-}
-
-NPUDeviceMem* GetDeviceMemories(c10::DeviceIndex deviceid) {
-  c10_npu::NPUGuard guard(deviceid);
-  size_t device_free;
-  size_t device_total;
-  NPU_CHECK_ERROR(aclrtGetMemInfo(ACL_HBM_MEM, &device_free, &device_total));
-  memory.totalGlobalMem = device_total;
-  memory.freeMem = device_free;
-  return &memory;
-}
-
-void BindGetDeviceMemories(PyObject* module) {
-  auto m = py::handle(module).cast<py::module>();
-  m.def(
-      "_npu_getDeviceMemories",
-      [](c10::DeviceIndex deviceid) -> NPUDeviceMem* { return GetDeviceMemories(deviceid); },
-      py::return_value_policy::reference);
-}
 
 void RegisterNpuPluggableAllocator(PyObject* module) {
   auto m = py::handle(module).cast<py::module>();
@@ -235,62 +146,6 @@ static PyObject* THNPModule_initExtension(PyObject* self, PyObject* noargs) {
   set_module_attr("default_generators", default_npu_generators);
 
   Py_RETURN_NONE;
-  END_HANDLE_TH_ERRORS
-}
-
-PyObject* THNPModule_npuSynchronize(PyObject* _unused, PyObject* noargs) {
-  HANDLE_TH_ERRORS
-  pybind11::gil_scoped_release no_gil;
-  c10_npu::npuSynchronizeDevice();
-  Py_RETURN_NONE;
-  END_HANDLE_TH_ERRORS
-}
-
-PyObject* THNPModule_setDevice_wrap(PyObject* self, PyObject* arg) {
-  HANDLE_TH_ERRORS
-  TORCH_CHECK(THPUtils_checkLong(arg), "invalid argument to setDevice");
-  {
-    pybind11::gil_scoped_release no_gil;
-    at::globalContext().lazyInitPrivateUse1();
-  }
-
-  auto device = THPUtils_unpackLong(arg);
-  c10_npu::set_device(static_cast<c10::DeviceIndex>(device));
-
-  Py_RETURN_NONE;
-  END_HANDLE_TH_ERRORS
-}
-
-PyObject* THNPModule_getDevice_wrap(PyObject* self, PyObject* noargs) {
-  HANDLE_TH_ERRORS
-  torch::utils::device_lazy_init(at::kPrivateUse1);
-  // NOLINTNEXTLINE(bugprone-signed-char-misuse)
-  auto device = static_cast<int32_t>(c10_npu::current_device());
-  return THPUtils_packInt32(device);
-  END_HANDLE_TH_ERRORS
-}
-
-PyObject* THNPModule_getDeviceCount_wrap(PyObject* self, PyObject* noargs) {
-  HANDLE_TH_ERRORS
-  return THPUtils_packUInt64(c10_npu::device_count());
-  END_HANDLE_TH_ERRORS
-}
-
-PyObject* THNPModule_npuCanDeviceAccessPeer_wrap(
-    PyObject* self,
-    PyObject* args) {
-  HANDLE_TH_ERRORS
-  PyObject* value_1 = nullptr;
-  PyObject* value_2 = nullptr;
-  if (!PyArg_ParseTuple(args, "OO", &value_1, &value_2)) {
-    throw torch::TypeError(
-        "Pybind failed to parse parameters." + PTA_ERROR(ErrCode::TYPE));
-  }
-  c10::DeviceIndex device_id = THPUtils_unpackDeviceIndex(value_1);
-  c10::DeviceIndex peer_device_id = THPUtils_unpackDeviceIndex(value_2);
-  auto can_access_peer =
-      c10_npu::acl::can_device_access_peer(device_id, peer_device_id);
-  return PyBool_FromLong(can_access_peer);
   END_HANDLE_TH_ERRORS
 }
 
@@ -994,23 +849,6 @@ static struct PyMethodDef THNPModule_methods[] = {
     {"_npu_set_run_yet_variable_to_false",
      (PyCFunction)THNPModule_set_run_yet_variable_to_false_wrap,
      METH_NOARGS,
-     nullptr},
-    {"_npu_synchronize",
-     (PyCFunction)THNPModule_npuSynchronize,
-     METH_NOARGS,
-     nullptr},
-    {"_npu_setDevice", (PyCFunction)THNPModule_setDevice_wrap, METH_O, nullptr},
-    {"_npu_getDevice",
-     (PyCFunction)THNPModule_getDevice_wrap,
-     METH_NOARGS,
-     nullptr},
-    {"_npu_getDeviceCount",
-     (PyCFunction)THNPModule_getDeviceCount_wrap,
-     METH_NOARGS,
-     nullptr},
-    {"_npu_canDeviceAccessPeer",
-     (PyCFunction)THNPModule_npuCanDeviceAccessPeer_wrap,
-     METH_VARARGS,
      nullptr},
     {"_npu_getCurrentStream",
      (PyCFunction)THNPModule_getCurrentStream_wrap,
