+++ torch_npu/csrc/npu/Module.h
@@ -2,14 +2,7 @@
 #define THNP_NPU_MODULE_INC
 #include "npu/core/NPUMacros.h"
 
-TORCH_NPU_API void RegisterNPUDeviceProperties(PyObject* module);
-TORCH_NPU_API void BindGetDeviceProperties(PyObject* module);
-TORCH_NPU_API void RegisterNPUDeviceMemories(PyObject* module);
-TORCH_NPU_API void BindGetDeviceMemories(PyObject* module);
 TORCH_NPU_API void RegisterNpuPluggableAllocator(PyObject* module);
-PyObject* THNPModule_getDevice_wrap(PyObject* self);
-PyObject* THNPModule_setDevice_wrap(PyObject* self, PyObject* arg);
-PyObject* THNPModule_getDeviceName_wrap(PyObject* self, PyObject* arg);
 PyObject* THNPModule_getDriverVersion(PyObject* self);
 PyObject* THNPModule_isDriverSufficient(PyObject* self);
 PyObject* THNPModule_getCurrentBlasHandle_wrap(PyObject* self);
